Use the New Type idiom for `Depth` and `Evaluation`.
Additionally, `Depth` is now an `i16` instead of a `u8` and `Evaluation`
is an `i32` instead of an `i16`. This allows for calculations with them
without having to worry about overflows.
The maximum and minimum for them remain the same though. This means that
when storing into the transposition table, they are converted easily
into the smaller, old, type (although they're still wrapped in a
struct).

Bench: 14916764

+++ build.rs
@@ -24,9 +24,8 @@ use std::{
     path::Path,
 };
 
-/// The difference between the root or leaf node (for height or depth
-/// respectively) and the current node.
-type Depth = u8;
+/// The difference between the leaf node and the current node.
+type Depth = i16;
 /// A table of base late move reductions.
 ///
 /// Indexed by the depth then number of legal moves.
